backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
requests==2.28.2
virtualenv==20.16.7
python-dotenv==1.0.0


